+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/ChatClient.kt
@@ -66,6 +66,9 @@ import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.client.notifications.ChatNotifications
 import io.getstream.chat.android.client.notifications.PushNotificationReceivedListener
 import io.getstream.chat.android.client.notifications.handler.ChatNotificationHandler
+import io.getstream.chat.android.client.notifications.handler.NotificationConfig
+import io.getstream.chat.android.client.notifications.handler.NotificationHandler
+import io.getstream.chat.android.client.notifications.handler.NotificationHandlerFactory
 import io.getstream.chat.android.client.notifications.storage.EncryptedPushNotificationsConfigStore
 import io.getstream.chat.android.client.notifications.storage.PushNotificationsConfig
 import io.getstream.chat.android.client.socket.ChatSocket
@@ -108,10 +111,6 @@ public class ChatClient internal constructor(
     private val userStateService: UserStateService = UserStateService(),
     private val tokenUtils: TokenUtils = TokenUtils,
 ) {
-
-    @InternalStreamChatApi
-    public val notificationHandler: ChatNotificationHandler = notifications.handler
-
     private var connectionListener: InitConnectionListener? = null
     private val logger = ChatLogger.get("Client")
     private val eventsObservable = ChatEventsObservable(socket, this)
@@ -1558,9 +1557,15 @@ public class ChatClient internal constructor(
         return "$header.$payload.$devSignature"
     }
 
+    /**
+     * Builder to initialize the singleton [ChatClient] instance and configure its parameters.
+     *
+     * @param apiKey The API key of your Stream Chat app obtained from the [Stream Dashboard](https://dashboard.getstream.io/).
+     * @param appContext The application [Context].
+     */
     public class Builder(private val apiKey: String, private val appContext: Context) : ChatClientBuilder() {
 
-        private var baseUrl: String = "chat-us-east-1.stream-io-api.com"
+        private var baseUrl: String = "chat.stream-io-api.com"
         private var cdnUrl: String = baseUrl
         private var baseTimeout = 30_000L
         private var cdnTimeout = 30_000L
@@ -1568,27 +1573,73 @@ public class ChatClient internal constructor(
         private var warmUp: Boolean = true
         private var callbackExecutor: Executor? = null
         private var loggerHandler: ChatLoggerHandler? = null
-        private var notificationsHandler: ChatNotificationHandler =
-            ChatNotificationHandler(appContext)
+        private var notificationsHandler: NotificationHandler? = null
+        private var notificationConfig: NotificationConfig = NotificationConfig()
         private var fileUploader: FileUploader? = null
         private val tokenManager: TokenManager = TokenManagerImpl()
         private var customOkHttpClient: OkHttpClient? = null
 
+        /**
+         * Sets the log level to be used by the client.
+         *
+         * See [ChatLogLevel] for details about the available options.
+         *
+         * We strongly recommend using [ChatLogLevel.NOTHING] in production builds,
+         * which produces no logs.
+         *
+         * @param level The log level to use.
+         */
         public fun logLevel(level: ChatLogLevel): Builder {
             logLevel = level
             return this
         }
 
+        /**
+         * Sets a [ChatLoggerHandler] instance that will receive log events from the SDK.
+         *
+         * Use this to forward SDK events to your own logging solutions.
+         *
+         * See the FirebaseLogger class in the UI Components sample app for an example implementation.
+         *
+         * @param loggerHandler Your custom [ChatLoggerHandler] implementation.
+         */
         public fun loggerHandler(loggerHandler: ChatLoggerHandler): Builder {
             this.loggerHandler = loggerHandler
             return this
         }
 
-        public fun notifications(notificationsHandler: ChatNotificationHandler): Builder {
+        /**
+         * Sets a custom [ChatNotificationHandler] that the SDK will use to handle everything
+         * around push notifications. Create your own subclass and override methods to customize
+         * notification appearance and behavior.
+         *
+         * See the [Push Notifications](https://staging.getstream.io/chat/docs/sdk/android/client/guides/push-notifications/)
+         * documentation for more information.
+         *
+         *
+         * @param notificationConfig Config push notification.
+         * @param notificationsHandler Your custom subclass of [ChatNotificationHandler].
+         */
+        @JvmOverloads
+        public fun notifications(
+            notificationConfig: NotificationConfig,
+            notificationsHandler: NotificationHandler = NotificationHandlerFactory.createNotificationHandler(context = appContext),
+        ): Builder = apply {
+            this.notificationConfig = notificationConfig
             this.notificationsHandler = notificationsHandler
-            return this
         }
 
+        /**
+         * Sets a custom file uploader implementation that will be used by the client
+         * to upload files and images.
+         *
+         * The default implementation uses Stream's own CDN to store these files,
+         * which has a 20 MB upload size limit.
+         *
+         * For more info, see [the File Uploads documentation](https://getstream.io/chat/docs/android/file_uploads/?language=kotlin).
+         *
+         * @param fileUploader Your custom implementation of [FileUploader].
+         */
         public fun fileUploader(fileUploader: FileUploader): Builder {
             this.fileUploader = fileUploader
             return this
@@ -1606,6 +1657,13 @@ public class ChatClient internal constructor(
             return this
         }
 
+        /**
+         * By default, ChatClient performs a dummy HTTP call to the Stream API
+         * when a user is set to initialize the HTTP connection and make subsequent
+         * requests reusing this connection execute faster.
+         *
+         * Calling this method disables this connection warm-up behavior.
+         */
         public fun disableWarmUp(): Builder = apply {
             warmUp = false
         }
@@ -1623,6 +1681,18 @@ public class ChatClient internal constructor(
             this.customOkHttpClient = okHttpClient
         }
 
+        /**
+         * Sets the base URL to be used by the client.
+         *
+         * By default, this is the URL of Stream's [Edge API Infrastructure](https://getstream.io/blog/chat-edge-infrastructure/),
+         * which provides low latency regardless of which region your Stream
+         * app is hosted in.
+         *
+         * You should only change this URL if you're on dedicated Stream
+         * Chat infrastructure.
+         *
+         * @param value The base URL to use.
+         */
         public fun baseUrl(value: String): Builder {
             var baseUrl = value
             if (baseUrl.startsWith("https://")) {
@@ -1638,6 +1708,11 @@ public class ChatClient internal constructor(
             return this
         }
 
+        @Deprecated(
+            message = "Do not use this method for file upload customization. Instead, implement the FileUploader interface and use the fileUploader method of this builder.",
+            level = DeprecationLevel.ERROR,
+            replaceWith = ReplaceWith("this.fileUploader(CustomFileUploader())")
+        )
         public fun cdnUrl(value: String): Builder {
             var cdnUrl = value
             if (cdnUrl.startsWith("https://")) {
@@ -1680,11 +1755,12 @@ public class ChatClient internal constructor(
                 ChatModule(
                     appContext,
                     config,
-                    notificationsHandler,
+                    notificationsHandler ?: NotificationHandlerFactory.createNotificationHandler(appContext),
+                    notificationConfig,
                     fileUploader,
                     tokenManager,
                     callbackExecutor,
-                    customOkHttpClient
+                    customOkHttpClient,
                 )
 
             val result = ChatClient(
@@ -1703,7 +1779,10 @@ public class ChatClient internal constructor(
     }
 
     public abstract class ChatClientBuilder @InternalStreamChatApi public constructor() {
-
+        /**
+         * Create a [ChatClient] instance based on the current configuration
+         * of the [Builder].
+         */
         public fun build(): ChatClient = buildChatClient().also {
             instance = it
         }
@@ -1771,11 +1850,6 @@ public class ChatClient internal constructor(
         @Throws(IllegalStateException::class)
         public fun dismissChannelNotifications(channelType: String, channelId: String) {
             ensureClientInitialized().notifications.dismissChannelNotifications(channelType, channelId)
-        }
-
-        @Throws(IllegalStateException::class)
-        internal fun dismissNotification(notificationId: Int) {
-            ensureClientInitialized().notifications.onDismissNotification(notificationId)
         }
 
         /**