+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/receivers/NotificationMessageReceiver.kt
@@ -18,7 +18,6 @@ internal class NotificationMessageReceiver : BroadcastReceiver() {
         private const val ACTION_READ = "com.getstream.sdk.chat.READ"
         private const val ACTION_REPLY = "com.getstream.sdk.chat.REPLY"
         private const val ACTION_DISMISS = "com.getstream.sdk.chat.DISMISS"
-        private const val KEY_NOTIFICATION_ID = "notification_id"
         private const val KEY_MESSAGE_ID = "message_id"
         private const val KEY_CHANNEL_ID = "id"
         private const val KEY_CHANNEL_TYPE = "type"
@@ -46,10 +45,7 @@ internal class NotificationMessageReceiver : BroadcastReceiver() {
             PendingIntent.getBroadcast(
                 context,
                 notificationId,
-                createNotifyIntent(context, notificationId, ACTION_REPLY).apply {
-                    putExtra(KEY_CHANNEL_ID, channel.id)
-                    putExtra(KEY_CHANNEL_TYPE, channel.type)
-                },
+                createNotifyIntent(context, channel, ACTION_REPLY),
                 MUTABLE_PENDING_INTENT_FLAGS,
             )
 
@@ -61,9 +57,7 @@ internal class NotificationMessageReceiver : BroadcastReceiver() {
         ): PendingIntent = PendingIntent.getBroadcast(
             context,
             notificationId,
-            createNotifyIntent(context, notificationId, ACTION_READ).apply {
-                putExtra(KEY_CHANNEL_ID, channel.id)
-                putExtra(KEY_CHANNEL_TYPE, channel.type)
+            createNotifyIntent(context, channel, ACTION_READ).apply {
                 putExtra(KEY_MESSAGE_ID, message.id)
             },
             IMMUTABLE_PENDING_INTENT_FLAGS,
@@ -72,10 +66,11 @@ internal class NotificationMessageReceiver : BroadcastReceiver() {
         internal fun createDismissPendingIntent(
             context: Context,
             notificationId: Int,
+            channel: Channel,
         ): PendingIntent = PendingIntent.getBroadcast(
             context,
             notificationId,
-            createNotifyIntent(context, notificationId, ACTION_DISMISS),
+            createNotifyIntent(context, channel, ACTION_DISMISS),
             IMMUTABLE_PENDING_INTENT_FLAGS,
         )
 
@@ -111,56 +106,52 @@ internal class NotificationMessageReceiver : BroadcastReceiver() {
                 .build()
         }
 
-        private fun createNotifyIntent(context: Context, notificationId: Int, action: String) =
+        private fun createNotifyIntent(context: Context, channel: Channel, action: String) =
             Intent(context, NotificationMessageReceiver::class.java).apply {
-                putExtra(KEY_NOTIFICATION_ID, notificationId)
+                putExtra(KEY_CHANNEL_ID, channel.id)
+                putExtra(KEY_CHANNEL_TYPE, channel.type)
                 this.action = action
             }
     }
 
-    override fun onReceive(context: Context?, intent: Intent?) {
-        when (intent?.action) {
-            ACTION_READ -> markAsRead(
-                intent.getStringExtra(KEY_MESSAGE_ID),
-                intent.getStringExtra(KEY_CHANNEL_ID),
-                intent.getStringExtra(KEY_CHANNEL_TYPE)
-            )
+    override fun onReceive(context: Context, intent: Intent) {
+        val channelType = intent.getStringExtra(KEY_CHANNEL_TYPE) ?: return
+        val channelId = intent.getStringExtra(KEY_CHANNEL_ID) ?: return
+        when (intent.action) {
+            ACTION_READ -> intent.getStringExtra(KEY_MESSAGE_ID)?.let { messageId ->
+                markAsRead(
+                    messageId,
+                    channelId,
+                    channelType,
+                )
+            }
             ACTION_REPLY -> {
-                val results = RemoteInput.getResultsFromIntent(intent)
-                if (results != null) {
+                val results = RemoteInput.getResultsFromIntent(intent).getCharSequence(KEY_TEXT_REPLY)?.let { message ->
                     replyText(
-                        intent.getStringExtra(KEY_CHANNEL_ID),
-                        intent.getStringExtra(KEY_CHANNEL_TYPE),
-                        results.getCharSequence(KEY_TEXT_REPLY)
+                        channelId,
+                        channelType,
+                        message,
                     )
                 }
             }
         }
-        intent?.getIntExtra(KEY_NOTIFICATION_ID, 0)?.let(::cancelNotification)
+        cancelNotification(channelType, channelId)
     }
 
-    private fun markAsRead(messageId: String?, channelId: String?, channelType: String?) {
+    private fun markAsRead(messageId: String, channelId: String, channelType: String) {
         if (!ChatClient.isInitialized) return
-
-        if (messageId.isNullOrBlank() || channelId.isNullOrBlank() || channelType.isNullOrBlank()) {
-            return
-        }
 
         ChatClient.instance().markMessageRead(channelType, channelId, messageId).enqueue()
     }
 
     private fun replyText(
-        channelId: String?,
-        type: String?,
-        messageChars: CharSequence?,
+        channelId: String,
+        type: String,
+        messageChars: CharSequence,
     ) {
         if (!ChatClient.isInitialized) return
 
-        val currentUser = ChatClient.instance().getCurrentUser()
-
-        if (messageChars.isNullOrBlank() || channelId.isNullOrBlank() || type.isNullOrBlank() || currentUser == null) {
-            return
-        }
+        val currentUser = ChatClient.instance().getCurrentUser() ?: return
 
         ChatClient.instance().sendMessage(
             channelType = type,
@@ -172,7 +163,7 @@ internal class NotificationMessageReceiver : BroadcastReceiver() {
         ).enqueue()
     }
 
-    private fun cancelNotification(notificationId: Int) {
-        ChatClient.dismissNotification(notificationId)
+    private fun cancelNotification(channelType: String, channelId: String) {
+        ChatClient.dismissChannelNotifications(channelType, channelId)
     }
 }