+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/event/EventHandlerImpl.kt
@@ -83,8 +83,6 @@ internal class EventHandlerImpl(
                 }
                 is ConnectedEvent -> {
                     logger.logI("Received ConnectedEvent, marking the domain as online and initialized")
-                    val recovered = domainImpl.isInitialized()
-
                     domainImpl.setOnline()
                     domainImpl.setInitialized()
                     domainImpl.scope.launch {
@@ -451,5 +449,9 @@ internal class EventHandlerImpl(
         } else {
             // for events of current user we keep "ownReactions" from the event
         }
+    }
+
+    internal fun clear() {
+        firstConnect = true
     }
 }