+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/message/attachment/AttachmentUploader.kt
@@ -17,8 +17,7 @@ internal class AttachmentUploader(
         channelType: String,
         channelId: String,
         attachment: Attachment,
-        attachmentTransformer: ((at: Attachment, file: File) -> Attachment)? = null,
-        progressCallback: ProgressCallback? = null
+        progressCallback: ProgressCallback? = null,
     ): Result<Attachment> {
         val file = checkNotNull(attachment.upload) { "An attachment needs to have a non null attachment.upload value" }
 
@@ -37,14 +36,7 @@ internal class AttachmentUploader(
                 mimeType = mimeType ?: "",
                 attachmentType = attachmentType,
                 url = result.data()
-            ).let {
-                // allow the user to change the format of the attachment
-                if (attachmentTransformer != null) {
-                    attachmentTransformer(it, file)
-                } else {
-                    it
-                }
-            }
+            )
             augmentedAttachment.uploadState = Attachment.UploadState.Success
             progressCallback?.onSuccess(augmentedAttachment.url)
             Result(augmentedAttachment)