+++ CHANGELOG.md
@@ -21,8 +21,14 @@
 ### ⬆️ Improved
 
 ### ✅ Added
+- Push Notification uses `MessagingStyle` on devices with API Version 23+
+- Push Notification configuration has been simplified, check our [docs](https://getstream.io/chat/docs/sdk/android/client/guides/push-notifications/#customizing-push-notifications) to see how it works
+- `NotificationHandler` interface allows you to implement your own Push Notification logic show/remove notifications. It is the new interface you need to use if you were using `ChatNotificationHandler` previously
+- `NotificationHandlerFactory` help you to use our default `NotificationHandler` implementations
 
 ### ⚠️ Changed
+- Some properties of `NotificationConfig` has been deprecated, check our [DEPRECATIONS](https://github.com/GetStream/stream-chat-android/blob/main/DEPRECATIONS.md) section
+- `ChatNotificationhandler` class has been deprecated, you need to use `NotificationHandler` now. Check our [DEPRECATIONS](https://github.com/GetStream/stream-chat-android/blob/main/DEPRECATIONS.md) section.
 
 ### ❌ Removed
 
@@ -31,6 +37,8 @@
 - Fixed issues with Proguard stripping response classes incorrectly
 
 ### ⬆️ Improved
+- Added KDocs for `ChatClient.Builder` methods.
+- `ChatClient` now defaults to using the `https://chat.stream-io-api.com` base URL, using [Stream's Edge API Infrastructure](https://getstream.io/blog/chat-edge-infrastructure/) instead of connecting to a region-specific API. If you're not on a dedicated chat infrastructure, remove any region-specific base URL settings from the `ChatClient.Builder` to use Edge instead.
 
 ### ✅ Added
 - 🚨 Breaking change: A new `Idle` state is added to `Attachment.UploadState`.
@@ -39,6 +47,7 @@
 
 ### ⚠️ Changed
 - 🚨 Breaking change: `Attachment.UploadState.InProgress` now is data class having two fields, `bytesUploaded: Long` and `totalBytes: Long` instead of object.
+- Deprecated the `ChatClient.Builder#cdnUrl` method. To customize file uploads, set a custom `FileUploader` implementation instead. More info in the documentation: [Using Your Own CDN](https://getstream.io/chat/docs/android/file_uploads/?language=kotlin#using-your-own-cdn).
 
 ### ❌ Removed
 
@@ -55,6 +64,7 @@
 
 ## stream-chat-android-ui-common
 ### 🐞 Fixed
+- Fixed PDF attachments previews
 
 ### ⬆️ Improved
 
@@ -66,10 +76,15 @@
 
 ## stream-chat-android-ui-components
 ### 🐞 Fixed
-
+- Fixed bug related to scroll of messages.
+- Updating attachments view holder only when attachments have changed. This fixes a problem with reloading gifs when reactions are added or removed.
 ### ⬆️ Improved
+- Now it is possible to set a custom `LinearLayoutManager` to `MessageListView`, this can be used to change stack of messages or revert the layout.
+- Removed full screen loading view when loading more message items on the `SearchResultListView`.
 
 ### ✅ Added
+- Added `MessageListView::getRecyclerView` method which exposes the inner `RecyclerView` with message list items. 
+- Added `MessageListView::setUserReactionClickListener` method to set a listener used when a reaction left by a user is clicked on the message options overlay.
 
 ### ⚠️ Changed